requirements.txt
aioesphomeapi==24.6.2
aiohappyeyeballs==2.4.3
aiohttp==3.11.11
aiosignal==1.3.2
ajsonrpc==1.2.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argcomplete==3.5.1
asgiref==3.8.1
async_interrupt==1.2.0
attrs==23.1.0
Authlib==1.3.1
backoff==2.2.1
bcrypt==4.2.1
bitarray==2.9.3
bitstring==4.2.3
blinker==1.8.2
bottle==0.13.2
build==1.2.2.post1
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chacha20poly1305-reuseable==0.13.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.2.0
cryptography==43.0.0
defcon==0.10.3
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
ecdsa==0.19.0
esphome==2024.11.1
esphome-dashboard==20241120.0
esptool==4.7.0
fastapi==0.115.6
filelock==3.15.4
flatbuffers==24.12.23
fonttools==4.55.0
freetype-py==2.5.1
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.6.1
gflanguages==0.7.1
gitdb==4.0.11
GitPython==3.1.43
glyphsets==1.0.0
glyphsLib==6.9.5
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.64.1
grpcio-health-checking==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
icmplib==3.0.4
idna==3.6
ifaddr==0.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
intel-openmp==2021.4.0
intelhex==2.3.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kconfiglib==13.7.1
kubernetes==31.0.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
mkl==2021.4.0
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
nltk==3.8.1
noiseprotocol==0.3.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==0.28.0
openstep_plist==0.4.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.13
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
paho-mqtt==1.6.1
pandas==2.2.2
pillow==10.4.0
platformio==6.1.16
posthog==3.7.4
propcache==0.2.1
protobuf==5.29.2
puremagic==1.27
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pyelftools==0.31
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.2
reedsolo==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.16.0
semantic-version==2.10.0
sentence-transformers==3.0.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.41.3
streamlit==1.36.0
sympy==1.13.0
tabulate==0.9.0
tbb==2021.13.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4
tqdm==4.66.4
transformers==4.42.4
trio==0.23.2
trio-websocket==0.11.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ufoLib2==0.17.0
unicodedata2==15.1.0
urllib3==2.1.0
uvicorn==0.30.6
validators==0.28.3
voluptuous==0.14.2
watchdog==4.0.1
watchfiles==1.0.3
weaviate-client==4.6.5
websocket-client==1.8.0
websockets==14.1
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zeroconf==0.132.2
zipp==3.21.0
